+++ include/arion/arion.hpp
@@ -28,6 +28,7 @@
 #include <arion/unicorn/unicorn.h>
 #include <vector>
 #include <variant>
+#include <optional>
 
 namespace arion
 {
@@ -48,6 +49,9 @@ class ARION_EXPORT Arion : public std::enable_shared_from_this<Arion>
     std::weak_ptr<Arion> parent;
     std::vector<std::shared_ptr<Arion>> children;
     std::exception_ptr uc_exception = nullptr;
+    bool afl_mode = false;
+    std::vector<int> afl_signals;
+    bool hard_stop = false;
     bool running = false;
     bool sync = false;
     pid_t pid;
@@ -58,7 +62,7 @@ class ARION_EXPORT Arion : public std::enable_shared_from_this<Arion>
     void init_dynamic_program(std::shared_ptr<ElfParser> prog_parser);
     void init_static_program(std::shared_ptr<ElfParser> prog_parser);
     void close_engines();
-    bool run_current(bool multi_process, bool main_inst);
+    bool run_current(bool multi_process, bool main_inst, std::optional<arion::ADDR> start = std::nullopt, std::optional<arion::ADDR> end = std::nullopt);
     void run_children();
 
   public:
@@ -94,11 +98,15 @@ class ARION_EXPORT Arion : public std::enable_shared_from_this<Arion>
     static std::weak_ptr<Arion> ARION_EXPORT get_arion_instance(pid_t pid);
     ~Arion();
     void cleanup_process();
-    void ARION_EXPORT run();
-    void ARION_EXPORT stop();
+    void ARION_EXPORT run(std::optional<arion::ADDR> start = std::nullopt, std::optional<arion::ADDR> end = std::nullopt);
+    void ARION_EXPORT run_from(arion::ADDR start);
+    void ARION_EXPORT run_to(arion::ADDR end);
+    void ARION_EXPORT stop(bool hard_stop=true);
     void crash(std::exception_ptr exception);
     void ARION_EXPORT sync_threads();
     bool ARION_EXPORT is_running();
+    void init_afl_mode(std::vector<int> signals);
+    void stop_afl_mode();
     std::shared_ptr<Arion> copy();
     bool ARION_EXPORT has_parent();
     std::shared_ptr<Arion> ARION_EXPORT get_parent();