+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Fixed dependency issues when linking in DEV mode
 - Added config for Arion instance
 - Implemented some syscalls
+- Added UnicornAFL fuzzing mode
+- Run from/to address
 
 # 1.0.1-alpha
 - Bumped CMake requirements